requirements.txt
py7zr
pillow
pystyle